Support for multiple effects maybe working
Didn't push this because it's not working on Mac; turns out I may have found a driver bug, so pushing now to test on Windows.

System has one big block of particles, create_particles() requests a block to be generated on the fly and removed when the effect expires.

Sending over an origin and timer for each block as instance attributes. May need a different implementation to be cross-platform